class for locarStorage work
